created checkout page
the check out page is for the customers only, they have to put in their name email and credit card info. the employees will not have to since they are just taking orders. Additionally, changed it so that employees will not see the favorites since they are just taking orders